add subcommands resize and swap
Added two subcommands, resize and swap.

Resize takes a vector of positive integer percentage values and resizes
all top-level containers according to said values from left-to-right.

Swap takes a direction (either left or right) and swaps the
currently-focused top-level container with the container either to the
left or to the right.