fix resize
Resize was not taking into account the fact that growing/shrinking a
container can only work if there is sufficient amounts of space to
steal from the adjacent container.

It's always possible to grow/shrink at least a container such that
one of the boundaries sits at the correct place, so just repeat this
until all of the containers have been resized.

Note that it's still possible for this to break as i3 container
percentages are floats, while grow/shrink must take integers.